pyproject.toml
[tool.poetry]
name = "repl_python3_TikTokmastermind"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyTikTokAPI = "^0.0.5"
pyTelegramBotAPI = "^4.2.2"
click8 = "^8.0.1"
fake-useragent = "^0.1.11"
selenium = "^4.1.0"
pytz = "^2021.3"
tzlocal = "^4.1"
click = "^8.0.3"
keyring = "^23.4.0"
aiogram = "^2.18"
pandas = "^1.3.5"
dash = "^2.0.0"
finnhub-python = "^2.4.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
